Unified Message data class
