+++ src/Data/Message/MessageDataClass.ts
@@ -1,6 +1,15 @@
-import { localStorageMessageDataClass } from './LocalStorage/localStorageMessageDataClass'
-import { pisaMessageDataClass } from './Pisa/pisaMessageDataClass'
+import { provideDataClass } from 'scrivito'
+import { pisaConfig } from '../pisaClient'
 
-export const Message = import.meta.env.ENABLE_PISA
-  ? pisaMessageDataClass()
-  : localStorageMessageDataClass()
+export const Message = provideDataClass(
+  'Message',
+  (async () => {
+    const restApi = await pisaConfig('message')
+
+    if (!restApi) {
+      return (await import('./messageParamsFallback')).messageParamsFallback()
+    }
+
+    return { restApi }
+  })(),
+)
