--- src/Data/Message/Pisa/pisaMessageDataClass.ts
@@ -1,6 +1,0 @@
-import { provideDataClass } from 'scrivito'
-import { pisaConfig } from '../../pisaClient'
-
-export function pisaMessageDataClass() {
-  return provideDataClass('Message', { restApi: pisaConfig('message') })
-}
